+++ docker/scripts/prod/entrypoint.sh
@@ -1,6 +1,12 @@
 #!/usr/bin/env bash
 set -e
 
+config=/config/recyclarr.yml
+if [[ "$RECYCLARR_CREATE_CONFIG" = true && ! -f "$config" ]]; then
+  echo "Creating default recyclarr.yml file..."
+  recyclarr create-config
+fi
+
 # If the script has any arguments, invoke the CLI instead
 if [ "$#" -gt 0 ]; then
     recyclarr "$@"